Created Formatter module and moved HTML and LaTeX to it
