+++ lib/iruby.rb
@@ -11,7 +11,8 @@ require 'iruby/kernel'
 require 'iruby/backend'
 require 'iruby/ostream'
 require 'iruby/input'
-require 'iruby/formatter'
+require 'iruby/formatter/latex'
+require 'iruby/formatter/html'
 require 'iruby/utils'
 require 'iruby/display'
 require 'iruby/comm'
